fix: remove unused lint import and sync with main
